refactor(api): move email into features
