+++ src/Infrastructure/CsvDownloadMiddleware.cs
@@ -1,5 +1,5 @@
 using app.Models;
-using app.Services;
+using app.Features.Email;
 using CsvHelper;
 using Dapper;
 using Microsoft.AspNetCore.Http;
